% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.funs.R
\name{bearing.xy}
\alias{bearing.xy}
\title{compute the heading between to cratesian points (radians counter clockwise from north or vertical)}
\usage{
bearing.xy(x1, x2, y1, y2)
}
\arguments{
\item{x1}{Numeric; point A x value}

\item{x2}{numeric; point B x value}

\item{y1}{numeric; point A y value}

\item{y2}{numeric; point B y value}
}
\value{
A single value in radians
}
\description{
compute the heading between to cratesian points (radians counter clockwise from north or vertical)
}
