% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{contrast.img}
\alias{contrast.img}
\title{Change image contrast}
\usage{
contrast.img(img = NULL, c = 2, out.dir = NULL, type = NULL)
}
\arguments{
\item{img}{character, a path to an image file or directory containing images.}

\item{c}{numeric, the factor by which to change the contrast. 'c'>1 enhances, 'c'<1 reduces.}

\item{out.dir}{character, the directory to which the image will be saved.}

\item{type}{character, image type (optional). Must be: "jpeg", "png", or "tiff (case ignored)." If missing, file format is automatically determined by file name extension(s).}
}
\description{
Changes the contrast of an image or images
}
\examples{


#retrieve image in system
 y <-system.file("extdata/img", "sunfish_BCF.jpg", package = "trackter")
od <- paste0(tempdir(),"/cropimg")
dir.create(od)

#display original
EBImage::display(EBImage::readImage(y),method="raster")


# crop and save as original format
contrast.img(img=y,c=0.5,out.dir=od)

#display modified image
EBImage::display(EBImage::readImage(paste0(od,"/sunfish_BCF.jpg")),method="raster")

#clean up
unlink(od,recursive=TRUE)
}
