% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.funs.R
\name{cosine.ang}
\alias{cosine.ang}
\title{Computes angle between two segments sharing a point.}
\usage{
cosine.ang(l, r, o)
}
\arguments{
\item{l}{Numeric; length of segment to the left}

\item{r}{Numeric; length of segment to the right}

\item{o}{Numeric; length of opposite segment}
}
\value{
A single value of the angle in radians
}
\description{
Computes angle between two segments of a triangle using law of cosines.
}
\examples{
#a right triangle
L=3
R=3
O=sqrt(L^2+R^2)
cosine.ang(L,R,O)
}
