% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin.2d.R
\name{cyclize}
\alias{cyclize}
\title{Calculate waveform cycles}
\usage{
cyclize(x, y, start = "peak", ...)
}
\arguments{
\item{x}{numeric; x position}

\item{y}{numeric; y position}

\item{start}{character; where the cycles should start either 'peak' or `trough'. See Details.}

\item{...}{additional arguments to be passed to \code{\link{features}}}
}
\value{
A a data table including the following: 

\itemize{
\item \code{cyc}: The cycle number.
\item \code{pos}: the position/index values of the x data (not the x values themselves).
}
}
\description{
Finds discrete cycles of wave-like data using critical points (peaks and troughs).
}
\details{
If "'start='peak'", the cycles start at critical points with high values, If "'start='trough'", cycles start at critical points with low values.
}
\examples{
library(ggplot2)
n <- 200
d <- data.frame(
x <- sort(runif(n)),
y <- exp(-0.2 * sin(10*pi*x)) + rnorm(n, sd=0.05)
)
#start at peaks
c <- cyclize(d$x,d$y)
d$cyc <- c$cyc
qplot(d=d,x,y,col=as.factor(cyc))

 #start at troughs
c <- cyclize(d$x,d$y,start="trough")
d$cyc <- c$cyc
qplot(data=d,x,y,col=as.factor(cyc))

}
\seealso{
\code{\link{features}}
}
