% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{data.overlay}
\alias{data.overlay}
\title{Plot data over an image}
\usage{
data.overlay(img, x, ...)
}
\arguments{
\item{img}{character, an image file path.}

\item{x}{a data frame, data table, or matrix with two columns, the first representing x and the second y coordinates.}

\item{...}{other arguments to be passed to \code{\link{points}}}
}
\value{
The image is plotted in the graphics device with points given in \code{x}
}
\description{
Simple wrapper for \code{\link{points}} to plot data derived from an image over that image.
}
\details{
Simply plots 2D dimensional data over the image specified in \code{img}. May be useful in plotting output from \code{trackter}'s kin functions.
}
\examples{

f <-system.file("extdata/img", "sunfish_BCF.jpg", package = "trackter")

d <- dim(EBImage::readImage(f))

x <- runif(10,1,d[1])
y <- runif(10,1,d[2])
pts <- cbind(x,y)

data.overlay(img=f,x=pts,col="red",type="p")
}
\seealso{
\code{\link{kin.free}}, \code{\link{kin.search}}, \code{\link{kin.simple}}
}
