% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.funs.R
\name{deg}
\alias{deg}
\title{Converts radians to degrees}
\usage{
deg(x)
}
\arguments{
\item{x}{Numeric; value in radians}
}
\value{
A single value
}
\description{
Converts radians to degrees
}
\seealso{
\code{\link{rad}}
}
