% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.funs.R
\name{dist.2d.line}
\alias{dist.2d.line}
\title{Computes orthogonal distance between a point and a line}
\usage{
dist.2d.line(x = NULL, y = NULL, slope = NULL, intercept = NULL, form = NULL)
}
\arguments{
\item{x}{numeric, the x coordinate of the point}

\item{y}{numeric, the y coordinate of the point}

\item{slope}{numeric, the slope of the line}

\item{intercept}{numeric, the slope of the line}

\item{form}{formula of type \code{lm} describing the line, ignored if \code{slope} and \code{intercept} are specified}
}
\value{
The distance between the point and line
}
\description{
Computes 2D orthogonal distance between a point and a line given the points coordinates and the line's slope and intercept or model formula
}
\details{
if \code{slope} and \code{intercept} are missing, a model formula of form \code{lm(y~x)} can be passed to \code{form}.
}
\examples{
x <- runif(1:10)
y <- x*0.5

xy.lm <- lm(y~x)

x.p <- y.p <- -3

#one way . . .
dist.2d.line(x.p,y.p,form=xy.lm)

#or another
dist.2d.line(x.p,y.p,slope=coef(xy.lm)[2],intercept=coef(xy.lm)[1])


}
