% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{gg.overlay}
\alias{gg.overlay}
\title{Plot output from \code{kin} functions}
\usage{
gg.overlay(
  kin = NULL,
  frames = NULL,
  under = "cont",
  over = "midline",
  zoom = FALSE,
  animate = FALSE,
  shadow = TRUE,
  alpha = 0.1,
  fps = 10,
  save = FALSE,
  filename = NULL,
  out.dir = NULL,
  ...
)
}
\arguments{
\item{kin}{a list of data returned from \code{kin} functions. See Details.}

\item{frames}{integer, a vector for frames within the data (a column common to all \code{kin} functions' output). Will subset the data or, if NULL, will not. If more than one frame is specified, plot will be faceted by frame with \code{facet_wrap} or animated across frames.}

\item{under}{character, the data to plot under \code{over}, the contour(s) from \code{kin} functions, i.e, 'cont' or 'cont.sm'. See Details.}

\item{over}{character, the data to plot over the contour(s). Must be either 'kin.dat', 'midline', 'both', 'fin', 'fin.pt', 'all', or 'none'. See Details.}

\item{zoom}{logical, should the plotted area zoom to the extents of the contours. See Detail.}

\item{animate}{logical, should an animation be plotted to the graphics device with state changes reflected by \code{frame}.}

\item{shadow}{logical,should \code{\link{shadow_mark}} be implemented to draw geometries in previous frames. Ignored if \code{animate=FALSE}.}

\item{alpha}{numeric, the opacity of geometries in previous frames during animation. Ignored if \code{animate=FALSE}.}

\item{fps}{integer, the play back speed of animation in frames per second. Ignored if \code{animate=FALSE}.}

\item{save}{logical, should the animation be saved as a GIF. Ignored if \code{animate=FALSE}. See Details.}

\item{filename}{character, the name given to the GIF file. Ignored if \code{animate=FALSE} or \code{save=FALSE}.}

\item{out.dir}{character, the file path of the directory to which the GIF file is saved. Ignored if \code{animate=FALSE} or \code{save=FALSE}.}

\item{...}{other arguments to be passed to midline and head point geometries, e.g. 'size', 'color'.}
}
\value{
A \code{ggplot} printed to the graphics device, a \code{gganimate} object plotted to graphics device, or a GIF file saved to a local directory specified in \code{out.dir}.
}
\description{
Plots data output from one of \code{trackter}'s \code{kin} functions using \code{ggplot2}. Useful for quickly visualizing results from \code{kin} functions. Data and geometries include body-contour polygons with midline points, head or fin points, both midline and head points, or none or all of these. Optionally builds and saves an animation of data with reference to frame from \code{kin} data.
}
\details{
Simply plots 2D dimensional data over contours retrieved from \code{kin} functions. May be useful in quickly assesses their results. The \code{under} layer can be one of the named contour data tables in lists returned by the \code{kin} functions \code{\link{kin.search}},\code{\link{kin.simple}}, \code{\link{kin.free}}, or \code{\link{fin.kin}}. The overlayed data layer specified by \code{over} are non-contour data from the \code{kin} functions.

If the list specified by 'kin' is from \code{\link{kin.search}},\code{\link{kin.simple}}, or  \code{\link{kin.free}}, 'over' must be 'cont' or 'cont.sm' and under one of 'midline','kin', 'none' or 'both'. 'midline' will produce the smoothed midline coordinates.

If the list specified by 'kin' is from \code{\link{fin.kin}}, 'over' must be 'cont' or 'comp' and under one of 'midline','fin', 'fin.pts', 'none' or 'all'.


Animations are produced with \code{gganimate} and optionally saved as a GIF with \code{anim_save} using the default \code{gifski_renderer()}.
}
\examples{

\dontrun{
#animate a ropefish swimming with its midline

#download example avi and place in subdirectory
f <- "https://github.com/ckenaley/exampledata/blob/master/ropefish.avi?raw=true"

download.file(f, paste0(tempdir(),"/ropefish.avi"))

dir.create(paste0(tempdir(),"/images"))

vid.to.images(paste0(tempdir(),"/ropefish.avi"), out.dir = paste0(tempdir(),"/images"))

kin <- kin.free(image.dir =paste0(tempdir(),"/images"),
par=TRUE,
      ml.smooth=list("spline",0.9),
      thr = "otsu",
      size.min=0.01,
      red=0.5
      )

gg.overlay(kin=kin,
frames=seq(20,320,20),
under="cont.sm",
over="midline",
size=1,
animate=TRUE,
col="red",
fps=10)

unlink(paste0(tempdir(),"/images"),recursive=TRUE)
}

}
\seealso{
\code{\link{geom_point}}, \code{\link{geom_polygon}},\code{\link{gganimate}}, \code{\link{kin.free}}, \code{\link{kin.search}}, \code{\link{kin.simple}},\code{\link{fin.kin}}
}
