% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg.R
\name{images.to.video}
\alias{images.to.video}
\title{Stitches images into a video file}
\usage{
images.to.video(
  image.dir = NULL,
  out.dir = NULL,
  vid.name = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{image.dir}{character; directory containing images to stitch.}

\item{out.dir}{character; directory in which to store video.}

\item{vid.name}{character; file name given to video including extension.}

\item{overwrite}{logical; should path described by \code{vid.name} be overwritten if it exists.}

\item{...}{other arguments to be passed to \code{av::av_encode_video}.}
}
\value{
Outputs a video of name "video.name+vid.ext".
}
\description{
Uses ffmpeg through the \code{av} package to stitch images into a video.
}
\details{
Assumes images are appended with a numeric sequence.
}
\examples{

#make some images
\donttest{
dir.create(paste0(tempdir(),"/images")) #make a directory to store images

a <- 2
b <- 3
theta <- seq(0,10*pi,0.01)
r <- a + b*theta
df <- data.frame(x=r*cos(theta), y=r*sin(theta)) # Cartesian coords
every.i <- 30
for(i in seq(1,length(theta),30)) {
  jpeg(paste0(tempdir(),"/images/image_",sprintf("\%03d",which(i==seq(1,length(theta),30))),".jpg"))
  with(df[1:i,],plot(x,y,xlim=range(df$x),ylim=range(df$y),col="red"))
  dev.off()
  }

images.to.video(image.dir=paste0(tempdir(),"/images"),vid.name="spiral.mp4",out.dir=tempdir())

file.exists(paste0(tempdir(),"/spiral.mp4"))

#clean up
unlink(paste0(tempdir(),"/spiral.mp4"))
unlink(paste0(tempdir(),"/images"),recursive=TRUE)
}
}
\seealso{
\code{\link{vid.to.images}}, \code{\link[av]{encoding}} \code{\link[av]{av_video_images}}.
}
