% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpf.R
\name{lpf}
\alias{lpf}
\title{Apply a low-pass filter to some time-series data.}
\usage{
lpf(y, t, f)
}
\arguments{
\item{y}{time series values.}

\item{t}{time.}

\item{f}{frequency,}
}
\value{
Filtered values that represent a rolling average, more or less...
}
\description{
Apply a low-pass filter to some time-series data.
}
\examples{
y <- sin(seq(1,30,0.1))
y <- sin(seq(1,30,0.1))
y <- jitter(y,2,2)
plot(y)
points(lpf(y,1,1),col="red",pch=3)
}
