% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{pick.pts}
\alias{pick.pts}
\title{2D Manual digitization of image data}
\usage{
pick.pts(img, n = 1000, review = TRUE, smoothing = NULL, smooth = NULL)
}
\arguments{
\item{img}{character, a one-dimension vector of image file paths.}

\item{n}{numeric, the max number of points to digitize for each frame (i.e., image file) in \code{img}. The number of points added to each frame may be less than this value. See details.}

\item{review}{logical. If 'TRUE', the user inputs 'return/enter' after points have been added for each frame. If 'FALSE', the function advances to the next frame without review after the last point has been added.}

\item{smoothing}{character, the smoothing method: either 'loess' or 'spline' for line smoothing or 'outline' for contours. See details.}

\item{smooth}{numeric; if \code{smoothing} is set to 'loess', this values is passed to 'span' parameter of \code{\link{loess}}. If \code{smoothing} is set to 'spline', this value is ignored and the smooth points are determined by the default values of \code{\link{spline}}. See details.}
}
\value{
A list of one or two data tables, alwyas \code{pts} with the picked points and, optionally, \code{smooth.pts} containing the the smoothed points if \code{smoothing} is chosen. Each contains the following columns:

\itemize{
\item 'frame', the frame file name
\item 'frame.n', the frame number
\item 'pt', the point number
\item x,y coordinates of the selected points
}
}
\description{
Uses the active graphics device for interactive digitization of shape data.
}
\details{
Users interact with each frame using the \code{\link{locator}} function. Although the maximum number of points for each frame is determined by \code{n}, the user may advance to the next frame by double-clicking on the last point. If \code{review=TRUE}, the user is asked to review the points and, if satisfied, a "return"/"enter" key stroke will advance the graphics device to the next frame. After double clicking in the last frame and a "return"/"enter" key stroke if \code{review=TRUE}, the results are returned. If \code{smooth=NULL}, only the clicked points are returned. If \code{smooth="spline"} or \code{smooth="loess"}, interpolated smoothed points are returned.

Smoothed points for outlines are produced with \code{\link{coo_smooth}}. Lines are smoothed with \code{\link{smooth.spline}} with \code{smooth} passed to the \code{spar} parameter or with \code{\link{smooth.spline}} with \code{smooth} passed to the \code{spar}
}
\examples{

\dontrun{
#download example avi video
f <- "https://github.com/ckenaley/exampledata/blob/master/sunfish_pect.avi?raw=true"
download.file(f,"sunfish.avi")

#extract images with ffmpeg operations and reduce them to 600 px wide with a filter
filt.red <- " -vf scale=600:-1 " #filter
dir.create(paste0(tempdir(),"/out"))
vid.to.images2(vid.path="sunfish.avi",filt = filt.red,out.dir = paste0(tempdir(),"/out")) 
y <- list.files(paste0(tempdir(),"/out"),full.names = TRUE)[1:3]
pick.pts(y,review=F)
}
}
\seealso{
\code{\link{locator}}, \code{\link{loess}}, \code{\link{smooth.spline}}
}
