% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.funs.R
\name{point.ang.orig}
\alias{point.ang.orig}
\title{Computes new position of a point rotating about an origin}
\usage{
point.ang.orig(p, o, theta)
}
\arguments{
\item{p}{numeric vector of length 2, the x and y coordinates of the point that will rotate}

\item{o}{numeric vector of length 2, the x and y coordinates of the origin}

\item{theta}{numeric, the angle of rotation}
}
\value{
A vector of length 2 containing the resulting x,y coordinates
}
\description{
Computes the coordinate position of a point after rotation about an origin.
}
\examples{
p <- c(2,2)
o <- c(0,0)

plot(c(-4,4),col=NULL)
points(p[1],p[2],col="red")
points(o[1],o[2])

new.p <- point.ang.orig(p,o,pi/2*-1)
points(new.p[1],new.p[2],col="blue")

}
