% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.funs.R
\name{rad}
\alias{rad}
\title{convert degrees to radians}
\usage{
rad(x)
}
\arguments{
\item{x}{Numeric; value in degrees}
}
\value{
A single value
}
\description{
convert degrees to radians
}
\seealso{
\code{\link{deg}}
}
