% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg2.R
\name{vid.to.images}
\alias{vid.to.images}
\title{Extracts images from a video file with ffmep}
\usage{
vid.to.images(vid.path = NULL, qual = 50)
}
\arguments{
\item{vid.path}{Character; path of video file to be processed.}

\item{qual}{numeric; the quality of the jpeg images to be rendered from 1-100\%. Defaults to 50\%.}
}
\value{
Extracts all the images of the video and saves them to an "images" directory with appended number sequence
}
\description{
Extracts images from a video file with ffmep
}
\examples{
#make a video with animation package
fun <- function(){
y <- sin(1:50)
x <- 1:50
for(i in 1:50) {
  plot(x[i],y[i],col="red",xlim=c(0,50),ylim=range(y))
  animation::ani.pause()
  }
}
saveVideo(fun(),video.name="wave.mp4",interval = 0.2)

vid.to.images(vid.path="wave.mp4",qual=100)

#see the images in the "images" directory
list.files( paste0(getwd(),"/","images"))
}
\seealso{
\code{\link{images.to.videos}}
}
