% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin.2d.R
\name{wave}
\alias{wave}
\title{Compute wavelength from the last half wave of sine-like waveform}
\usage{
wave(x, y, p = F)
}
\arguments{
\item{x}{Numeric; x position}

\item{y}{numeric; y position}

\item{p}{logical; should a plot of the last half wavelength}
}
\value{
a list with half wavelength "l" and position of the crest, "p"
}
\description{
Compute wavelength from the last half wave of sine-like waveform
}
\examples{

#Find length of the last half wave
x <- seq(0,pi,0.1)
y <- sin(x*pi)

wave(x,y,p=T)
}
