% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin.2d.R
\name{amp.freq}
\alias{amp.freq}
\title{Computes amplitude and frequency of wave-like data}
\usage{
amp.freq(x = NULL, y, sf = 100)
}
\arguments{
\item{x}{Numeric; x position (or sample number)}

\item{y}{numeric; y position}

\item{sf}{numeric; sample frequency (i.e., how often was x and y sampled) in Hz}
}
\value{
a list with of the following:
\itemize{
\item \code{a}: a numeric vector of amplitudes ('a').
\item \code{pos}: a numeric vector of the position of critical points.
\item \code{f}: a numeric vector of the peak-to-peak frequencies (if the first critical point is a peak) or trough-to-trough frequencies (if the first critical point is a trough).
\item \code{snr}:a numeric value of the signal to noise ratio.
}
}
\description{
Computes amplitude(s) and frequencies(s) of a wave form, amongst other things, based on a sampling frequency
}
\details{
computes amplitudes as half the distance between each successive pair of critical points (i.e., points where the derivative is zero).
}
\examples{

#Compute waveform patterns
x <- seq(0,pi,0.1)
y <- sin(x^1.3*pi)
plot(x,y)

amp.freq(x=x,y=y)

}
\seealso{
\code{\link{features}}
}
