% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.funs.R
\name{bearing.xy}
\alias{bearing.xy}
\title{Compute the heading between to cartesian points}
\usage{
bearing.xy(x1, x2, y1, y2)
}
\arguments{
\item{x1}{Numeric; point A x value}

\item{x2}{numeric; point B x value}

\item{y1}{numeric; point A y value}

\item{y2}{numeric; point B y value}
}
\value{
A single value in radians

a single value in radians
}
\description{
Computes the heading (radians counter clockwise from north or vertical)
}
\examples{
#example
A <- c(0,0)
B <- c(-5,5)
thet <- bearing.xy(A[1],B[1],A[2],B[2])
deg(thet)


}
