% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg2.R
\name{images.to.video}
\alias{images.to.video}
\title{Stitches images into an AVI file}
\usage{
images.to.video(image.dir = NULL, vid.name = NULL, qual = 50,
  vid.ext = ".mp4", frame.rate = 10)
}
\arguments{
\item{image.dir}{Character; directory containing images to stich.}

\item{vid.name}{character; file name to be give video.}

\item{qual}{numeric; the quality of the video rendered from 1-100\%. Defaults to 50\%.}

\item{vid.ext}{chacracter; video type to output. mp4 currently works best.}

\item{frame.rate}{numeric; video frame rate in fps.}
}
\value{
Outputs a video of name "video.name+vid.ext".
}
\description{
Stitches images into an AVI file
}
\details{
Assumes images are appended with a numeric sequence.
}
\examples{

#make some images
require(emojifont)

y <- sin(1:50)
x <- 1:50
dir.create("images") #make a directory to stor images
for(i in 1:50) {
  jpeg(paste0(getwd(),"/images/image",sprintf("\%03d",i),".jpg"))
  plot(x[i],y[i],col="red",xlim=c(0,50),ylim=range(y),cex=0)
  text(x[i], y[i], labels=emoji('cow'), cex=6, col='steelblue', family='EmojiOne')
  dev.off()
  }

images.to.video(image.dir=paste0(getwd(),"/images"),vid.name="flyingcow",frame.rate=5,qual=100)

}
\seealso{
\code{\link{vid.to.images}}
}
